requirements.txt
langgraph==0.2.19
requests==2.31.0
beautifulsoup4==4.12.2
html2text==2020.1.16
lxml==4.9.3
tqdm==4.66.1
python-dotenv==1.0.0

# Step 2 - Gemini RAG
google-generativeai==0.7.2
chromadb==0.4.24
markdown==3.6
numpy==1.26.4
